requirements.txt
flask
gunicorn
psycopg2-binary
pyJWT
pytest
dataclasses